refactored language objects nesting
