--- src/constants/translations/en/modals.json
@@ -1,7 +1,0 @@
-{
-  "emailConfirm": "Your email has been successfully verified!",
-  "emailNotConfirm": "Your email address has not been verified!",
-  "emailAlreadyConfirm": "Your email address has already been verified!",
-  "emailReject-badToken-description": "The confirmation is either invalid or has expired. Please sign up again and confirm your email address.",
-  "emailReject-alreadyConfirmed-description": "User email has been already confirmed"
-}
