+++ src/constants/translations/ua/emailModals.json
@@ -1,0 +1,9 @@
+{
+  "emailConfirm": "Ваш email успішно перевірено!",
+  "emailNotConfirm": "Ваш email не перевірено!",
+  "emailAlreadyConfirm": "Ваш email вже перевірено!",
+  "emailReject" : {
+    "badToken" : "Підтвердження або недійсне, або термін його дії минув. Будь ласка, зареєструйтеся ще раз і підтвердьте свою електронну адресу.",
+    "alreadyConfirmed" : "Електронну адресу користувача вже підтверджено."
+  }
+}
