+++ src/constants/translations/ua/index.js
@@ -5,7 +5,7 @@ import header from './header.json'
 import questions from './questions.json'
 import titles from './titles.json'
 import login from './login.json'
-import modals from './modals.json'
+import modals from './emailModals.json'
 
 const ua = {
   translations: { common, errorPage, button, header, questions, titles, login, modals }
